backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
openai>=1.44.0
sse-starlette>=1.6.1
python-dotenv>=1.0.1
asyncpg>=0.29.0

